+++ frappe_s3_attachment/frappe_s3_attachment/doctype/s3_file_attachment/s3_file_attachment.js
@@ -6,17 +6,10 @@ frappe.ui.form.on('S3 File Attachment', {
 
 	},
 	migrate_existing_files: function (frm) {
-        frappe.msgprint("Local files getting migrated", "S3 Migration");
-        frappe.call({
-            method: "frappe_s3_attachment.controller.migrate_existing_files",
-            callback: function (data) {
-                if (data.message) {
-					frappe.msgprint('Upload Successful')
-					location.reload(true);
-                } else {
-                    frappe.msgprint('Retry');
-                }
-            }
-        });
-    },
+		frappe.call({
+			method: "frappe_s3_attachment.controller.migrate_existing_files",
+			freeze: true,
+			freeze_message: __("Queuing S3 migration..."),
+		});
+	},
 });